refactor: move utility functions in separate module
